evaluate returns changed_file_paths

+++ lib/sashimi_tanpopo/file_updater.rb
@@ -38,12 +38,17 @@ module SashimiTanpopo
       # @return [Hash<Symbol, String>]
       attr_reader :params
 
+      # @!attribute [r] changed_file_paths
+      # @return [Array<String>]
+      attr_reader :changed_file_paths
+
       # @param params [Hash<Symbol, String>]
       # @param dry_run [Boolean]
       # @param is_colored [Boolean] Whether show color diff
       def initialize(params:, dry_run:, is_colored:)
         @params = params
         @dry_run = dry_run
+        @changed_file_paths = []
 
         @diffy_format = is_colored ? :color : :text
       end
@@ -53,7 +58,9 @@ module SashimiTanpopo
       # @yieldparam content [String] content of file
       def update_file(pattern, &block)
         Dir.glob(pattern).each do |path|
-          update_single_file(path, &block)
+          is_changed = update_single_file(path, &block)
+
+          @changed_file_paths << path if is_changed
         end
       end
 
@@ -62,21 +69,23 @@ module SashimiTanpopo
       # @param path [String]
       # @param block [Proc]
       # @yieldparam content [String] content of file
+      #
+      # @return [Boolean] Whether file is changed
       def update_single_file(path, &block)
-        return unless File.exist?(path)
+        return false unless File.exist?(path)
 
         content = File.read(path)
 
         result = yield content.dup
 
         # File isn't changed
-        return if content == result
+        return false if content == result
 
         show_diff(content, result)
 
-        return if @dry_run
+        File.write(path, result) unless @dry_run
 
-        File.write(path, result)
+        true
       end
 
       # @param str1 [String]
@@ -104,6 +113,7 @@ module SashimiTanpopo
               eval(#{recipe_body.dump}, nil, #{recipe_path.dump}, 1)
             end
           end
+          @context.changed_file_paths
         RUBY
 
         @target_dir = target_dir
